Renaming common.h in DEC
