+++ examples/dec/DECExamplesCommon.h
