+++ examples/dec/exampleDiscreteExteriorCalculusSolve.cpp
@@ -3,7 +3,7 @@
 
 #include <QApplication>
 
-#include "common.h"
+#include "DECExamplesCommon.h"
 
 // always include EigenSupport.h before any other Eigen headers
 #include "DGtal/math/linalg/EigenSupport.h"
