+++ tests/dec/DECCommon.h
